refactor: relation migrations
